API

API

CMBLensing.Cℓ_to_cov — Method.

Convert power spectrum Cℓ to a flat sky diagonal covariance

source

CMBLensing.MAP_joint — Method.

MAP_joint(ds::DataSet; L=LenseFlow, Nϕ=nothing, quasi_sample=nothing, nsteps=10, Ncg=500, cgtol=1e-1, αtol=1e-5, αmax=0.5, progress=false)

Compute the maximum a posteri estimate (MAP) from the joint posterior (can also do a quasi-sample).

The ds argument stores the data and other relevant objects for the dataset being considered. L gives which type of lensing operator to use.

ϕstart can be used to specify the starting point of the minimizer, but this is not necessary and otherwise it will start at ϕ=0.

Nϕ can optionally specify an estimate of the ϕ effective noise, and if provided is used to estimate a Hessian which is used in the ϕ quasi-Newton-Rhapson step. Nϕ=:qe automatically uses the quadratic estimator noise.

This function can also be used to draw quasi-samples, wherein for the f step, we draw a sample from P(f|ϕ) instead of maximizing it (ie instead of computing Wiener filter). quasi_sample can be set to an integer seed, in which case each time in the f step we draw a same-seeded sample. If quasi_sample is instead just true, then each iteration in the algorithm draws a different sample so the solution bounces around rather than asymptoting to a maximum.

The following arguments control the maximiation procedure, and can generally be left at their defaults:

nsteps - The number of iteration steps to do (each iteration updates f then updates ϕ)

Ncg - Maximum number of conjugate gradient steps during the f update

cgtol - Conjugrate gradient tolerance (will stop at cgtol or Ncg, whichever is first)

αtol - Tolerance for the linesearch in the ϕ quasi-Newton-Rhapson step, x′ = x - α*H⁻¹*g

αmax - Maximum value for α in the linesearch

progress - Whether to print out conjugate gradient progress.

Returns a tuple (f, ϕ, tr) where f is the best-fit (or quasi-sample) field, ϕ is the lensing potential, and tr contains info about the run.

source

CMBLensing.MAP_marg — Method.

MAP_marg( ds; kwargs...)

Compute the maximum a posteri estimate (MAP) of the marginl posterior.

source

CMBLensing.animate — Method.

animate(fields::Vector{\<:Vector{\<:Field}}; interval=50, motionblur=true, kwargs...)

source

CMBLensing.bcggd — Method.

Branching ConjugateGradient-GradientDescent

Algortihm: In parallel do: a) 2 steps of gradient descent with Ncg steps of conjugate gradient towards H⁻¹g at each iteration b) 1 steps of gradient descent with β*Ncg steps of conjugate gradient towards H⁻¹g

Take whichever is a lower lnP. If we take (b), then update Ncg = β*Ncg. 

β is chosen such that (a) and (b) take roughly the same ammount of time to
compute (usually β≈2 but could be higher if there's alot of overhead for a GD step)

Arguments:

t : which parametrization to use (i.e. t=0 for unlensed or t=1 for lensed)

fₜϕ_start : fₜϕ to start at

Nsteps : how many iterations of the above algorithm to do

Ncg : the starting value of Ncg

β : scaling factor described above

L : Lensing operator to use for gradient descent

LJ : Lensing operator to use for the Hessian calculation

source

CMBLensing.cg — Function.

Simple generic conjugate gradient implementation that works on Vectors, Fields, etc...

source

CMBLensing.gdsteps — Method.

Do Ngd gradient descent steps, with Ncg conjugate gradient steps towards solving H⁻¹g at each iteration.

Arguments:

t : which parametrization to use (i.e. t=0 for unlensed or t=1 for lensed)

fₜϕ_cur : fₜϕ to start at

Ngd, Ncg : number of gradient descent and conjugate gradient steps

L : Lensing operator to use for gradient descent

LJ : Lensing operator to use for the Hessian calculation

Returns:

lnP(fₜϕ), fₜϕ, trace

source

CMBLensing.lensing_wiener_filter — Method.

lensing_wiener_filter(ds::DataSet, L, which=:wf)

Computes either,

the Wiener filter at fixed $\phi$, i.e. the best-fit of

\[\mathcal{P}(f\,|\,\phi,d)\]

a sample from $\mathcal{P}(f\,|\,\phi,d)$

The data model assumed is,

\[d = \mathcal{M} \mathcal{B} \mathcal{L} \, f + n\]

Note that the noise is defined as un-debeamed and also unmasked (so it needs to be filled in outside the mask if using real data). The mask, $\mathcal{M}$, can be any composition of real and/or fourier space diagonal operators.

The argument ds::DataSet stores the mask, $\mathcal{M}$, the beam/instrumental transfer functions, $\mathcal{B}$, as well as the various covariances which are needed.

The which parameter controls which operation to do and can be one of three things:

:wf - Compute the Wiener filter

:sample - Compute a sample from the posterior

:fluctuation - Compute a fluctuation around the mean (i.e. a sample minus the Wiener filter)

source

CMBLensing.load_sim_dataset — Method.

load_sim_dataset

Create a DataSet object with some simulated data.

source

CMBLensing.noisecls — Method.

μKarcminT: temperature noise in μK-arcmin

beamFWHM: beam-FWHM in arcmin

source

CMBLensing.norm² — Method.

norm²(f::Field, L::LinOp)

Shorthand for f⋅(L\f), i.e. the squared-norm of f w.r.t. the operator L.

source

CMBLensing.pcg — Function.

Preconditioned conjugate gradient. P should be symmetric and roughly √A⁻¹

source

CMBLensing.pcg2 — Function.

pcg2(M, A, b, x=M\b; nsteps=length(b), tol=sqrt(eps()), progress=false, callback=nothing, hist=nothing, histmod=1)

Compute x = A\b (where A is positive definite) by conjugate gradient. M is the preconditioner and should approximate A, and M \ x should be fast.

The solver will stop either after nsteps iterations or when dot(r,r)<tol (where r=A*x-b is the residual at that step), whichever occurs first.

Info from the iterations of the solver can be returned if hist is specified. hist can be one or a tuple of:

:i - current iteration number

:x - current solution

:r - current residual r=A*x-b

:res - the norm of r

:t - the time elapsed (in seconds) since the start of the algorithm

histmod can be used to include every N-th iteration only in hist.

source

CMBLensing.pixwin — Method.

pixwin(θpix, ℓ)

Returns the pixel window function for square flat-sky pixels of width θpix (in arcmin) evaluated at some ℓs.

The pixel window function is defined so that if you start with white noise at infinitely high resolution and pixelize it down a resolution θpix, its power spectrum will be given by pixwin(θpix, ℓ)^2.

source

CMBLensing.resimulate — Method.

resimulate(ds::DataSet; f=..., ϕ=...)

Resimulate the data in a given dataset, potentially at a fixed f and/or ϕ (both are resimulate if not provided)

source

CMBLensing.sample_joint — Method.

sample_joint(ds::DataSet; kwargs...)

Sample from the joint PDF of P(f,ϕ,r). Runs nworkers() chains in parallel using pmap.

Possible keyword arguments:

nsamps_per_chain - the number of samples per chain

nchunk - do nchunk steps in-between parallel chain communication

nsavemaps - save maps into chain every nsavemaps steps

nburnin_always_accept - the first nburnin_always_accept steps, always accept HMC steps independent of integration error

nburnin_fixθ - the first nburnin_fixθ steps, fix θ at its starting point

chains - resume an existing chain (starts a new one if nothing)

θrange - range and density to grid sample parameters as a NamedTuple, e.g. (Aϕ=range(0.7,1.3,length=20),).

θstart - starting values of parameters as a NamedTuple, e.g. (Aϕ=1.2,), or nothing to randomly sample from θrange

ϕstart - starting ϕ as a Field, or :quasi_sample or :best_fit

source

CMBLensing.sptlike_mask — Method.

An SPT-like mask, based mostly on,

For the boundary:

"[First we] calculate the distance from the nearest masked pixel, and this distance map is smoothed using a Gaussian beam of FWHM = 15'. This smoothing is applied to soften the corners of the mask. The distance map is then used to apodize the binary mask with a Gaussian beam of FWHM = 30'" - Omori et al. 2017

and for point sources:

"The SPTpol deep field contains a number of bright point sources, which we mask. We identify all sources detected at > 5σ at 150 GHz from Mocanu et al. (2013), which corresponds to a flux-cut of approximately 6 mJy; all pixels within 5' of each source are masked in both the TOD filtering and the final maps. We extend this mask to 10' for all very bright sources detected at > 75σ. We also use a 10' radius to mask all galaxy clusters de- tected in this field by Vanderlinde et al. (2010) using the Sunyaev-Zel’dovich effect. This masking removes ap- proximately 120 sources, cutting 5 deg² of the field. We additionally multiply the maps by a sky mask that down- weights the noisy edges of the maps." Story et al. 2015

Note: some of the above numbers are very slightly tweaked.

source

CMBLensing.symplectic_integrate — Method.

symplectic_integrate(x₀, p₀, Λ, U, δUδx, N=50, ϵ=0.1, progress=false)

Do a symplectic integration of the potential energy U (with gradient δUδx) starting from point x₀ with momentum p₀ and mass matrix Λ. The number of steps is N and the step size ϵ.

Returns ΔH, xᵢ, pᵢ corresponding to change in Hamiltonian, and final position and momenta. If hist is specified a trace of requested variables throughout each step is also returned.

source

CMBLensing.ud_grade — Method.

ud_grade(f::Field, θnew, mode=:map, deconv_pixwin=true, anti_aliasing=true)

Up- or down-grades field f to new resolution θnew (only in integer steps). Two modes are available specified by the mode argument:

*`:map`     : Up/downgrade by replicating/averaging pixels in map-space
*`:fourier` : Up/downgrade by extending/truncating the Fourier grid

For :map mode, two additional options are possible. If deconv_pixwin is true, deconvolves the pixel window function from the downgraded map so the spectrum of the new and old maps are the same. If anti_aliasing is true, filters out frequencies above Nyquist prior to down-sampling.

source

CMBLensing.ϕqe — Function.

ϕqe(d, Cf, Cf̃, Cn, Cϕ=nothing)

Compute quadratic estimate for ϕ given data.

d - data

Cf, Cf̃ - unlensed and lensed beamed theory covariances

Cn - noise covariance (beam not deconvolved)

Cϕ - (optional) lensing potential theory covariance. if provided, the result is Wiener filtered, otherwise the unbiased estimate is retured.

Returns a tuple of (ϕqe, Nϕ) where ϕqe is the quadratic estimate and Nϕ is the N0 noise bias.

source

PyPlot.plot — Method.

plot(f::Field; kwargs...)
plot(fs::VecOrMat{\<:Field}; kwarg...)

Plotting fields.

source

CMBLensing.Mnyq — Method.

filter out the single row/column in the real FFT matrix M which corresponds to exactly the nyquist frequency

source

CMBLensing.antilensing — Method.

Create from an existing PowerLens operator one that lenses by -ϕ instead.

source

CMBLensing.camb_cl_files — Method.

camb_cl_files(filename_root)

Loads Cℓ's from some CAMB output files. filename_root should be such that $(filenameroot)scalCls.dat, etc... are the CAMB data files.

source

CMBLensing.fftsyms — Method.

Arguments m and n refer to the sizes of an m×n matrix (call it A) that is the output of a real FFT (thus m=n÷2+1)

Returns a tuple of (ireal, iimag, negks) where these are

* ireal - m×n mask corrsponding to unique real entries of A
* iimag - m×n mask corrsponding to unique imaginary entries of A
* negks - m×n matrix of giving the index into A where the negative k-vector
          is, s.t. A[i,j] = A[negks[i,j]]'

source

CMBLensing.fieldvalues — Method.

Return the type's fields as a tuple

source

CMBLensing.full — Method.

full(::Type{Fin}, ::Type{Fout}, L::LinOp; progress=true)
full(::Type{F}, L::LinOp; progress=true)

Construct an explicit matrix representation of the linear operator L by applying it to a set of vectors which form a complete basis. The Fin and Fout types should be fields which specify the input and output bases for the representation (or just F if L is square and we want the same input/output bases)

The name full is to be consistent with Julia's SparseArrays where full builds a full matrix from a sparse one.

source

CMBLensing.rfft2vec — Method.

Convert a matrix A which is the output of a real FFT to a real vector, keeping only unqiue real/imaginary entries of A

source

CMBLensing.unfold — Method.

Convert an M×N matrix (with M=N÷2+1) which is the output a real FFT to a full N×N one via symmetries.

source

CMBLensing.vec2rfft — Method.

Convert a vector produced by rfft2vec back into a complex matrix.

source

CMBLensing.Σ — Method.

Σ(ϕ, ds, ::Type{L}=LenseFlow) where {L}
Σ(L::LenseOp, ds)

Operator for the data covariance, Cn + PMBLCfL'B'M'P', which can applied and inverted.

source

CMBLensing.δf̃_δf — Method.

δf̃(f,ϕ)/δf * v

source

CMBLensing.δf̃_δfᴴ — Method.

(δf̃(f,ϕ)/δf)ᴴ * v

source

CMBLensing.δf̃_δϕ — Method.

δf̃(f,ϕ)/δϕ * v

source

CMBLensing.δf̃_δϕᴴ — Method.

(δf̃(f,ϕ)/δϕ)ᴴ * v

source

CMBLensing.δ²f̃_δϕ² — Method.

Compute wᴴ * δ²f̃(f,ϕ)/δϕ² * v

source

CMBLensing.jrk4 — Method.

Solve for y(t₁) with 4th order Runge-Kutta assuming dy/dt = F(t,y) and y(t₀) = y₀

Arguments

F! : a function F!(v,t,y) which sets v=F(t,y)

source