API

API

Main.CMBLensing.MAP_joint — Method.

MAP_joint(ds::DataSet; L=LenseFlow, Nϕ=nothing, quasi_sample=nothing, nsteps=10, Ncg=500, cgtol=1e-1, αtol=1e-5, αmax=0.5, progress=false)

Compute the maximum a posteri estimate (MAP) from the joint posterior (can also do a quasi-sample).

The ds argument stores the data and other relevant objects for the dataset being considered. L gives which type of lensing operator to use.

ϕstart can be used to specify the starting point of the minimizer, but this is not necessary and otherwise it will start at ϕ=0.

Nϕ can optionally specify an estimate of the ϕ effective noise, and if provided is used to estimate a Hessian which is used in the ϕ quasi-Newton-Rhapson step. Nϕ=:qe automatically uses the quadratic estimator noise.

This function can also be used to draw quasi-samples, wherein for the f step, we draw a sample from P(f|ϕ) instead of maximizing it (ie instead of computing Wiener filter). quasi_sample can be set to an integer seed, in which case each time in the f step we draw a same-seeded sample. If quasi_sample is instead just true, then each iteration in the algorithm draws a different sample so the solution bounces around rather than asymptoting to a maximum.

The following arguments control the maximiation procedure, and can generally be left at their defaults:

nsteps - The number of iteration steps to do (each iteration updates f then updates ϕ)

Ncg - Maximum number of conjugate gradient steps during the f update

cgtol - Conjugrate gradient tolerance (will stop at cgtol or Ncg, whichever is first)

αtol - Tolerance for the linesearch in the ϕ quasi-Newton-Rhapson step, x′ = x - α*H⁻¹*g

αmax - Maximum value for α in the linesearch

progress - Whether to print out conjugate gradient progress.

Returns a tuple (f, ϕ, tr) where f is the best-fit (or quasi-sample) field, ϕ is the lensing potential, and tr contains info about the run.

Main.CMBLensing.MAP_marg — Method.

MAP_marg( ds; kwargs...)

Compute the maximum a posteri estimate (MAP) of the marginl posterior.

Main.CMBLensing.animate — Method.

animate(fields::Vector{\<:Vector{\<:Field}}; interval=50, motionblur=false, kwargs...)

Main.CMBLensing.argmaxf_lnP — Method.

argmaxf_lnP(ϕ,                ds::DataSet; kwargs...)
argmaxf_lnP(ϕ, θ::NamedTuple, ds::DataSet; kwargs...)
argmaxf_lnP(Lϕ::LenseOp,      ds::DataSet; kwargs...)

Computes either the Wiener filter at fixed $\phi$, or a sample from this slice along the posterior.

Keyword arguments:

which : :wf, :sample, or fluctuation to compute 1) the Wiener filter, i.e. the best-fit of $\mathcal{P}(f\,|\,\phi,d)$, 2) a sample from $\mathcal{P}(f\,|\,\phi,d)$, or 3) a sample minus the Wiener filter, i.e. the fluctuation on top of the mean.

guess : starting guess for f for the conjugate gradient solver

kwargs : all other arguments are passed to conjugate_gradient

Main.CMBLensing.gradhess — Method.

gradhess(f)

Compute the gradient gⁱ = ∇ⁱf, and the hessian, Hⁱⱼ = ∇ⱼ∇ⁱf

Main.CMBLensing.lnP — Function.

lnP(t, fₜ, ϕₜ,                ds::DataSet, Lϕ=nothing)
lnP(t, fₜ, ϕₜ, θ::NamedTuple, ds::DataSet, Lϕ=nothing)

Compute the log posterior probability in the joint parameterization as a function of the field, $f_t$, the lensing potential, $\phi_t$, and possibly some cosmological parameters, $\theta$. The subscript $t$ can refer to either a "time", e.g. passing t=0 corresponds to the unlensed parametrization and t=1 to the lensed one, or can be :mix correpsonding to the mixed parametrization. In all cases, the arguments fₜ and ϕₜ should then be $f$ and $\phi$ in that particular parametrization.

If any parameters $\theta$ are passed, we also include the three determinant terms to properly normalize the posterior.

The argument ds should be a DataSet and stores the masks, data, etc... needed to construct the posterior. If Lϕ is provided, it will be used as memory to recache the lensing operator at the specified ϕ.

Main.CMBLensing.load_sim_dataset — Method.

load_sim_dataset

Create a DataSet object with some simulated data. E.g.

@unpack f,ϕ,ds = load_sim_dataset(;
    θpix  = 2,
    Nside = 128,
    use   = :I,
    T     = Float32
);

Main.CMBLensing.mix — Method.

mix(f, ϕ, θ, ds)

Compute the mixed (f°, ϕ°) from the unlensed field f and lensing potential ϕ.

Main.CMBLensing.noiseCℓs — Method.

noiseCℓs(;μKarcminT, beamFWHM=0, ℓmax=8000, ℓknee=100, αknee=3)

Compute the (:TT,:EE,:BB,:TE) noise power spectra given white noise + 1/f. Polarization noise is scaled by $\sqrt{2}$ relative to μKarcminT. beamFWHM is in arcmin.

Main.CMBLensing.pixwin — Method.

pixwin(θpix, ℓ)

Returns the pixel window function for square flat-sky pixels of width θpix (in arcmin) evaluated at some ℓs. This is the scaling of k-modes, the scaling of the power spectrum will be pixwin^2.

Main.CMBLensing.quadratic_estimate — Method.

quadratic_estimate(ds::DataSet, which; wiener_filtered=true)
quadratic_estimate((ds1::DataSet,ds2::DataSet), which; wiener_filtered=true)

Compute quadratic estimate of ϕ given data.

The ds or (ds1,ds2) tuple contain the DataSet object(s) which houses the data and covariances used in the estimate. Note that only the Fourier-diagonal approximations for the beam, mask, and noise,, i.e. ds.B̂, ds.M̂, and ds.Cn̂, are accounted for. To account full operators (if they are not actually Fourier-diagonal), you should compute the impact using Monte Carlo.

If a tuple is passed in, the result will come from correlating the data from ds1 with that from ds2, which can be useful for debugging / isolating various noise terms.

An optional keyword argument AL can be passed in in case the QE normalization was already computed, in which case it won't be recomputed during the calculation.

Returns a NamedTuple (ϕqe, AL, Nϕ) where ϕqe is the (possibly Wiener filtered, depending on wiener_filtered option) quadratic estimate, AL is the normalization (which is already applied to ϕqe, it does not need to be applied again), and Nϕ is the analytic N0 noise bias (Nϕ==AL if using unlensed weights, currently only Nϕ==AL is always returned, no matter the weights)

Main.CMBLensing.resimulate — Method.

resimulate(ds::DataSet; f=..., ϕ=...)

Resimulate the data in a given dataset, potentially at a fixed f and/or ϕ (both are resimulated if not provided)

Main.CMBLensing.sample_joint — Method.

sample_joint(ds::DataSet; kwargs...)

Sample from the joint PDF of P(f,ϕ,θ). Runs nworkers() chains in parallel using pmap.

Possible keyword arguments:

nsamps_per_chain - the number of samples per chain

nchunk - do nchunk steps in-between parallel chain communication

nsavemaps - save maps into chain every nsavemaps steps

nburnin_always_accept - the first nburnin_always_accept steps, always accept HMC steps independent of integration error

nburnin_fixθ - the first nburnin_fixθ steps, fix θ at its starting point

chains - resume an existing chain (starts a new one if nothing)

θrange - range and density to grid sample parameters as a NamedTuple, e.g. (Aϕ=range(0.7,1.3,length=20),).

θstart - starting values of parameters as a NamedTuple, e.g. (Aϕ=1.2,), or nothing to randomly sample from θrange

ϕstart - starting ϕ as a Field, or :quasi_sample or :best_fit

metadata - does nothing, but is saved into the chain file

nhmc - the number of HMC passes per ϕ Gibbs step (default: 1)

Main.CMBLensing.symplectic_integrate — Method.

symplectic_integrate(x₀, p₀, Λ, U, δUδx, N=50, ϵ=0.1, progress=false)

Do a symplectic integration of the potential energy U (with gradient δUδx) starting from point x₀ with momentum p₀ and mass matrix Λ. The number of steps is N and the step size ϵ.

Returns ΔH, xᵢ, pᵢ corresponding to change in Hamiltonian, and final position and momenta. If hist is specified a trace of requested variables throughout each step is also returned.

Main.CMBLensing.ud_grade — Method.

ud_grade(f::Field, θnew, mode=:map, deconv_pixwin=true, anti_aliasing=true)

Up- or down-grades field f to new resolution θnew (only in integer steps). Two modes are available specified by the mode argument:

*`:map`     : Up/downgrade by replicating/averaging pixels in map-space
*`:fourier` : Up/downgrade by extending/truncating the Fourier grid

For :map mode, two additional options are possible. If deconv_pixwin is true, deconvolves the pixel window function from the downgraded map so the spectrum of the new and old maps are the same. If anti_aliasing is true, filters out frequencies above Nyquist prior to down-sampling.

Main.CMBLensing.unmix — Method.

unmix(f°, ϕ°, θ, ds)

Compute the unmixed/unlensed (f, ϕ) from the mixed field f° and mixed lensing potential ϕ°.

Main.CMBLensing.δlnP_δfϕₜ — Function.

δlnP_δfϕₜ(t, fₜ, ϕ,                ds, Lϕ=nothing)
δlnP_δfϕₜ(t, fₜ, ϕ, θ::NamedTuple, ds, Lϕ=nothing)

Compute a gradient of the log posterior probability. See lnP for definition of arguments of this function.

The return type is a FieldTuple corresponding to the $(f_t,\phi)$ derivative.

PyPlot.plot — Method.

plot(f::Field; kwargs...)
plot(fs::VecOrMat{\<:Field}; kwarg...)

Plotting fields.

Main.CMBLensing.FlatEBFourier — Type.

# main constructor:
FlatEBFourier(El::AbstractMatrix, Bl::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatEBFourier{P}(El::AbstractMatrix, Bl::AbstractMatrix) # specify pixelization P explicilty
FlatEBFourier{P,T}(El::AbstractMatrix, Bl::AbstractMatrix) # additionally, convert elements to type Complex{T}
FlatEBFourier{P,T,M<:AbstractMatrix{Complex{T}}}(El::M, Bl::M) # specify everything explicilty

Construct a FlatEBFourier object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatEBMap — Type.

# main constructor:
FlatEBMap(Ex::AbstractMatrix, Bx::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatEBMap{P}(Ex::AbstractMatrix, Bx::AbstractMatrix) # specify pixelization P explicilty
FlatEBMap{P,T}(Ex::AbstractMatrix, Bx::AbstractMatrix) # additionally, convert elements to type T
FlatEBMap{P,T,M<:AbstractMatrix{T}}(Ex::M, Bx::M) # specify everything explicilty

Construct a FlatEBMap object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatFourier — Type.

# main constructor:
FlatFourier(Il::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatFourier{P}(Il::AbstractMatrix) # specify pixelization P explicilty
FlatFourier{P,T}(Il::AbstractMatrix) # additionally, convert elements to type Complex{T}
FlatFourier{P,T,M<:AbstractMatrix{Complex{T}}}(Il::M) # specify everything explicilty

Construct a FlatFourier object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatIEBFourier — Type.

# main constructor:
FlatIEBFourier(Il::AbstractMatrix, El::AbstractMatrix, Bl::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatIEBFourier{P}(Il::AbstractMatrix, El::AbstractMatrix, Bl::AbstractMatrix) # specify pixelization P explicilty
FlatIEBFourier{P,T}(Il::AbstractMatrix, El::AbstractMatrix, Bl::AbstractMatrix) # additionally, convert elements to type Complex{T}
FlatIEBFourier{P,T,M<:AbstractMatrix{Complex{T}}}(Il::M, El::M, Bl::M) # specify everything explicilty

Construct a FlatIEBFourier object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatIEBMap — Type.

# main constructor:
FlatIEBMap(Ix::AbstractMatrix, Ex::AbstractMatrix, Bx::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatIEBMap{P}(Ix::AbstractMatrix, Ex::AbstractMatrix, Bx::AbstractMatrix) # specify pixelization P explicilty
FlatIEBMap{P,T}(Ix::AbstractMatrix, Ex::AbstractMatrix, Bx::AbstractMatrix) # additionally, convert elements to type T
FlatIEBMap{P,T,M<:AbstractMatrix{T}}(Ix::M, Ex::M, Bx::M) # specify everything explicilty

Construct a FlatIEBMap object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatIQUFourier — Type.

# main constructor:
FlatIQUFourier(Il::AbstractMatrix, Ql::AbstractMatrix, Ul::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatIQUFourier{P}(Il::AbstractMatrix, Ql::AbstractMatrix, Ul::AbstractMatrix) # specify pixelization P explicilty
FlatIQUFourier{P,T}(Il::AbstractMatrix, Ql::AbstractMatrix, Ul::AbstractMatrix) # additionally, convert elements to type Complex{T}
FlatIQUFourier{P,T,M<:AbstractMatrix{Complex{T}}}(Il::M, Ql::M, Ul::M) # specify everything explicilty

Construct a FlatIQUFourier object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatIQUMap — Type.

# main constructor:
FlatIQUMap(Ix::AbstractMatrix, Qx::AbstractMatrix, Ux::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatIQUMap{P}(Ix::AbstractMatrix, Qx::AbstractMatrix, Ux::AbstractMatrix) # specify pixelization P explicilty
FlatIQUMap{P,T}(Ix::AbstractMatrix, Qx::AbstractMatrix, Ux::AbstractMatrix) # additionally, convert elements to type T
FlatIQUMap{P,T,M<:AbstractMatrix{T}}(Ix::M, Qx::M, Ux::M) # specify everything explicilty

Construct a FlatIQUMap object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatMap — Type.

# main constructor:
FlatMap(Ix::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatMap{P}(Ix::AbstractMatrix) # specify pixelization P explicilty
FlatMap{P,T}(Ix::AbstractMatrix) # additionally, convert elements to type T
FlatMap{P,T,M<:AbstractMatrix{T}}(Ix::M) # specify everything explicilty

Construct a FlatMap object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatQUFourier — Type.

# main constructor:
FlatQUFourier(Ql::AbstractMatrix, Ul::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatQUFourier{P}(Ql::AbstractMatrix, Ul::AbstractMatrix) # specify pixelization P explicilty
FlatQUFourier{P,T}(Ql::AbstractMatrix, Ul::AbstractMatrix) # additionally, convert elements to type Complex{T}
FlatQUFourier{P,T,M<:AbstractMatrix{Complex{T}}}(Ql::M, Ul::M) # specify everything explicilty

Construct a FlatQUFourier object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.FlatQUMap — Type.

# main constructor:
FlatQUMap(Qx::AbstractMatrix, Ux::AbstractMatrix[, θpix={resolution in arcmin}, ∂mode={fourier∂ or map∂})

# more low-level:
FlatQUMap{P}(Qx::AbstractMatrix, Ux::AbstractMatrix) # specify pixelization P explicilty
FlatQUMap{P,T}(Qx::AbstractMatrix, Ux::AbstractMatrix) # additionally, convert elements to type T
FlatQUMap{P,T,M<:AbstractMatrix{T}}(Qx::M, Ux::M) # specify everything explicilty

Construct a FlatQUMap object. The top form of the constructor is most convenient for interactive work, while the others may be more useful for low-level code.

Main.CMBLensing.ParamDependentOp — Type.

ParamDependentOp(recompute_function::Function)
ParamDependentOp(recompute_function!::Function, mem)

Creates an ImplicitOp which depends on some parameters $\theta$ and can be evaluated at various values of these parameters. There are two forms to construct this operator. In the first form, recompute_function should be a function which accepts keyword arguments for $\theta$ and returns the operator. Each keyword must have a default value; the operator will act as if evaluated at these defaults unless it is explicitly evaluated at other parameters. In the second form, we can preallocate some memory for the results mem, in which case recompute_function! should additionally accept a single positional argument holding this memory, which should then be assigned in-place.

Example:

Cϕ₀ = Diagonal(...) # some fixed Diagonal operator
Cϕ = ParamDependentOp((;Aϕ=1)->Aϕ*Cϕ₀) # create ParamDependentOp

Cϕ(Aϕ=1.1) * ϕ   # Cϕ(Aϕ=1.1) is equal to 1.1*Cϕ₀
Cϕ * ϕ           # Cϕ alone will act like Cϕ(Aϕ=1) because that was the default above

# a version which preallocates the memory:
Cϕ = ParamDependentOp((mem;Aϕ=1)->(@. mem = Aϕ*Cϕ₀), similar(Cϕ₀))

Note: if you are doing parallel work, global variables referred to in the recompute_function need to be distributed to all workers. A more robust solution is to avoid globals entirely and instead ensure all variables are "closed" over (and hence will automatically get distributed). This will happen by default if defining the ParamDependentOp inside any function, or can be forced at the global scope by wrapping everything in a let-block, e.g.:

Cϕ = let Cϕ₀=Cϕ₀
    ParamDependentOp((;Aϕ=1)->Aϕ*Cϕ₀)
end

After executing the code above, Cϕ is now ready to be shipped to any workers and will work regardless of what global variables are defined on these workers.

Also note: if you want to use these ops in a DataSet which may be evaluated at other parameters, you can use a definition like e.g. (;Aϕ=1, _...)-> to capture only the parameter you care about while allowing others to exist.

Main.CMBLensing.RK4Solver — Method.

Solve for y(t₁) with 4th order Runge-Kutta assuming dy/dt = F(t,y) and y(t₀) = y₀

Arguments

F! : a function F!(v,t,y) which sets v=F(t,y)

Main.CMBLensing.@namedtuple — Macro.

Pack some variables into a NamedTuple

> x = 3
> y = 4
> @namedtuple(x, y, z=5)
(x=3,y=4,z=5)

Main.CMBLensing.@repeated — Macro.

Return a tuple with the expression repeated n times

Main.CMBLensing.antilensing — Method.

Create from an existing PowerLens operator one that lenses by -ϕ instead.

Main.CMBLensing.beamCℓs — Method.

beamCℓs(;beamFWHM, ℓmax=8000)

Compute the beam power spectrum, often called $W_\ell$. A map should be multiplied by $\sqrt$ of this.

Main.CMBLensing.conjugate_gradient — Function.

conjugate_gradient(M, A, b, x=M\b; nsteps=length(b), tol=sqrt(eps()), progress=false, callback=nothing, hist=nothing, histmod=1)

Compute x = A\b (where A is positive definite) by conjugate gradient. M is the preconditioner and should approximate A, and M \ x should be fast.

The solver will stop either after nsteps iterations or when dot(r,r)<tol (where r=A*x-b is the residual at that step), whichever occurs first.

Info from the iterations of the solver can be returned if hist is specified. hist can be one or a tuple of:

:i - current iteration number

:x - current solution

:r - current residual r=A*x-b

:res - the norm of r

:t - the time elapsed (in seconds) since the start of the algorithm

histmod can be used to include every N-th iteration only in hist.

Main.CMBLensing.fftsyms — Method.

Arguments m and n refer to the sizes of an m×n matrix (call it A) that is the output of a real FFT (thus m=n÷2+1)

Returns a tuple of (ireal, iimag, negks) where these are

* ireal - m×n mask corrsponding to unique real entries of A
* iimag - m×n mask corrsponding to unique imaginary entries of A
* negks - m×n matrix of giving the index into A where the negative k-vector
          is, s.t. A[i,j] = A[negks[i,j]]'

Main.CMBLensing.fieldvalues — Method.

Return the type's fields as a tuple

Main.CMBLensing.get_term_memoizer — Method.

All of the terms in the quadratic estimate and normalization expressions look like

C * l[i] * l̂[j] * l̂[k] * ...

where C is some field or diagonal covariance. For example, there's a term in the EB estimator that looks like:

(CE * (CẼ+Cn) \ d[:E])) * l[i] * l̂[j] * l̂[k]

(where note that l̂[j] and l̂[k] are unit vectors, but l[i] is not). The function get_term_memoizer returns a function term which could be called in the following way to compute this term:

term((CE * (CẼ+Cn) \ d[:E])), [i], j, k)

(note that the fact that l[i] is not a unit vector is specified by putting the [i] index in brackets).

Additionally, all of these terms are symmetric in their indices, i.e. in (i,j,k) in this case. The term function is smart about this, and is memoized so that each unique set of indices is only computed once. This leads to a pretty drastic speedup for terms with many indices like those that arize in the EE and EB normalizations, and lets us write code which is both clear and fast without having to think too hard about these symmetries.

Main.CMBLensing.rfft2vec — Method.

Convert a matrix A which is the output of a real FFT to a real vector, keeping only unqiue real/imaginary entries of A

Main.CMBLensing.unfold — Method.

Convert an M×N matrix (with M=N÷2+1) which is the output a real FFT to a full N×N one via symmetries.

Main.CMBLensing.vec2rfft — Method.

Convert a vector produced by rfft2vec back into a complex matrix.

Main.CMBLensing.Σ — Method.

Σ(ϕ, ds, ::Type{L}=LenseFlow) where {L}
Σ(L::LenseOp, ds)

An operator for the data covariance, Cn + PMBLCfL'B'M'P', which can applied and inverted.

Main.CMBLensing.@! — Macro.

Rewrites @! x = f(args...) to x = f!(x,args...)

Special cases for * and \ forward to mul! and ldiv!, respectively.

Main.CMBLensing.@dictpack — Macro.

Pack some variables in a dictionary

> x = 3
> y = 4
> @dictpack x y z=>5
Dict(:x=>3,:y=>4,:z=>5)

Main.CMBLensing.@invokelatest — Macro.

@invokelatest expr...

Rewrites all non-broadcasted function calls anywhere within an expression to use Base.invokelatest. This means functions can be called that have a newer world age, at the price of making things non-inferrable.

Main.CMBLensing.@ondemand — Macro.

@ondemand(Package.function)(args...; kwargs...)
@ondemand(Package.Submodule.function)(args...; kwargs...)

Just like calling Package.function or Package.Submodule.function, but Package will be loaded on-demand if it is not already loaded. The call is no longer inferrable.

Main.CMBLensing.@subst — Macro.

@subst sum(x*$(y+1) for x=1:2)

becomes

let tmp=(y+1)
    sum(x*tmp for x=1:2)
end

to aid in writing clear/succinct code that doesn't recompute things unnecessarily.

Main.CMBLensing.@sym_memo — Macro.

# symmetric in any of its final arguments except for bar:
@sym_memo foo(bar, @sym(args...)) = <body> 
# symmetric in (i,j), but not baz
@sym_memo foo(baz, @sym(i, j)) = <body>

The @sym_memo macro should be applied to a definition of a function which is symmetric in some of its arguments. The arguments in which its symmetric are specified by being wrapping them in @sym, and they must come at the very end. The resulting function will be memoized and permutations of the arguments which are equal due to symmetry will only be computed once.

Main.CMBLensing.FFTW_NUM_THREADS — Constant.

The number of FFTW threads to use. This must be set via e.g.:

CMBLensing.FFTW_NUM_THREADS[] = 4

before creating any FlatField objects; subsequent changes to this variable will be ignored. The default value is Sys.CPU_THREADS.